Moved files and debugged.
