+++ src/world/spawn.lua
@@ -1,11 +1,11 @@
 -- Spawn some ships or other objects based on a table containing all of the
 -- data required to produce them. This module is designed to be called in
 -- SceneParser and Saves.
-local PartRegistry = require("shipparts/partRegistry")
+local PartRegistry = require("world/shipparts/partRegistry")
 local Util = require("util")
-local Structure = require("structure")
+local Structure = require("world/structure")
 local GridTable = require("gridTable")
-local World = require("world")
+local World = require("world/world")
 
 local Spawn = {}
